+++ src/LogicalOROperation.java
@@ -1,17 +1,17 @@
 public class LogicalOROperation implements Operation{
-    private int bitwiseResult;
+    private double bitwiseResult;
 
     public LogicalOROperation() {
         bitwiseResult = 0;
     }
 
     @Override
-    public void performOperation(int operand1, int operand2) {
+    public void performOperation(double operand1, double operand2) {
         bitwiseResult = (operand1!=0) || (operand2!=0) ? 1:0;
     }
 
     @Override
-    public int getBitwiseResult() {
+    public double getBitwiseResult() {
         return bitwiseResult;
     }
 }
